backend/pyproject.toml
[tool.poetry]
name = "my-ai"
version = "0.1.0"
description = "my langchain based ai"
authors = ["Uyiosa <hiskonxeptz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
beautifulsoup4 = "4.12.2"
curl_cffi = "0.5.5"
fake_useragent = "1.1.3"
fastapi = "0.95.2"
llama_index = "0.6.8"
pycryptodome = "3.17"
pydantic = "1.10.5"
pypasser = "0.0.5"
requests = "2.28.2"
retrying = "1.3.4"
selenium = "4.9.1"
tenacity = "8.2.2"
tls_client = "0.2.1"
websocket_client = "1.5.1"
langchain = "^0.0.181"
google-api-python-client = "^2.86.0"
wikipedia = "^1.4.0"
lxml = "^4.9.2"
python-dotenv = "^1.0.0"
uvicorn = "^0.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

